Remove automatic synchronization in end_frame.
Callers now need to manually call `sync_all` when they wish to commit
the index to the disk. This drastically increases write throughput, but
transfers more responsibility to the caller. Overall the change is for
the better as now the caller no longer has to guess when the index is
automatically synced.