Allow configuration of the disk material in setup.
I'm not sure how to provide an Image handle for the texture here. Will
leave that for future work when we decide that we want to paint textures
on disks.

One option is to provide yet another type that describes the material
parameters again, but with a string for the asset. Then setup() could
load the asset from the server. Alternately, I suppose the caller could
add a first stage to the pipe that loads the image from the asset server
and passes it to the next.